refactor: split FindReplaceComponent
