+++ packages/remirror__react-components/src/find-replace/replace-controller.tsx
@@ -1,0 +1,18 @@
+import { Button, ButtonGroup } from '@mui/material';
+import React, { FC } from 'react';
+
+export const ReplaceController: FC<{
+  replace: () => void;
+  replaceAll: () => void;
+}> = ({ replace, replaceAll }) => {
+  return (
+    <ButtonGroup variant='outlined' size='small'>
+      <Button aria-label='Relace' sx={{ textTransform: 'none' }} onClick={replace}>
+        Replace
+      </Button>
+      <Button aria-label='Relace all' sx={{ textTransform: 'none' }} onClick={replaceAll}>
+        All
+      </Button>
+    </ButtonGroup>
+  );
+};
